+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### 9.1.5 (11/11/2024 - )
 
 * log-exporter: change gsutil to "gcloud storage"
+* log-exporter: remove trace exporting
+  > trace is only for troubleshooting, unnecessary for long term
 * mongo: update driver to 5.2.1
 
 > gsutil is deprecated and requires old version of python